// apps/desktop/src/bindings/proto/SessionPrompt.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "./serde_json/JsonValue";
import type { PromptState } from "./PromptState";
import type { UserId } from "./UserId";
import type { WritingSessionId } from "./WritingSessionId";

export type SessionPrompt = { id: number, session_id: WritingSessionId, creator_id: UserId, position: number, 
/**
 * TipTap JSON document.
 */
prompt_doc: JsonValue, timer_seconds: number | null, state: PromptState, started_at: number | null, ends_at: number | null, ended_at: number | null, };